% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{Ark}
\alias{Ark}
\title{A Class to represent an ark.}
\description{
A Class to represent an ark.

A Class to represent an ark.
}
\details{
An Ark object can create unique pseudonyms.
Given the same input, it will always return the same psuedonym.
No pseudonym will repeat.
}
\examples{

## ------------------------------------------------
## Method `Ark$print`
## ------------------------------------------------

ark <- Ark()
print(ark)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ark$new()}}
\item \href{#method-print}{\code{Ark$print()}}
\item \href{#method-length}{\code{Ark$length()}}
\item \href{#method-as.list}{\code{Ark$as.list()}}
\item \href{#method-pseudonymize}{\code{Ark$pseudonymize()}}
\item \href{#method-clone}{\code{Ark$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new arc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Ark} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print Ark object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ark <- Ark()
print(ark)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\if{latex}{\out{\hypertarget{method-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$length(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.list"></a>}}
\if{latex}{\out{\hypertarget{method-as.list}{}}}
\subsection{Method \code{as.list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$as.list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pseudonymize"></a>}}
\if{latex}{\out{\hypertarget{method-pseudonymize}{}}}
\subsection{Method \code{pseudonymize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$pseudonymize(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
