% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_permutation.R
\name{get_n_remaining}
\alias{get_n_remaining}
\title{Get the number of unused elements left in a random permutation}
\usage{
get_n_remaining(f)
}
\arguments{
\item{f}{Function created by \code{random_permutation()}}
}
\value{
Integer, number of elements left in the permutation
}
\description{
Get the number of unused elements left in a random permutation
}
\keyword{internal}
