% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind2subs}
\alias{ind2subs}
\title{Convert linear index to matrix subscripts}
\usage{
ind2subs(ind, dims)
}
\arguments{
\item{ind}{An integer vector with linear indexes}

\item{dims}{An n-dimensional integer vector. Each element of this vector
indicates the size of the corresponding dimension in the n-dimensional
matrixc.}
}
\value{
A list of n lists of equal length. Each list corresponds to one of
the n dimensions of the matrix. Rows indicate subscripts. Suitable for
use with mapping functions.
}
\description{
Takes a vector of integers (1D linear indexed) and converts them to matrix
subscripts in n-dimensional matrix. Modeled after MATLAB's \code{ind2sub()}
}
\keyword{internal}
