% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\title{Coalescing infix OR operator}
\description{
Operator that provides default value for NULL.
See \code{\link[rlang]{op-null-default}} for details.
}
\keyword{internal}
