% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_permutation.R
\name{remove_remaining}
\alias{remove_remaining}
\title{Remove numbers from a lazy non-repeating random number generator}
\usage{
remove_remaining(f, i)
}
\arguments{
\item{f}{Function created by \code{random_permutation()}}

\item{i}{Integer vector of numbers that should  be removed from remaining
numbers in the random permutation that is produced by f}
}
\value{
A random permutation function which will not produce the numbers in
i anymore.
}
\description{
Modify the random permutation function f in order to remove
numbers i from the numbers that are available in the random permutation
}
\keyword{internal}
