% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subs2ind}
\alias{subs2ind}
\title{Convert matrix subscripts to linear index}
\usage{
subs2ind(subs, dims)
}
\arguments{
\item{subs}{Subscripts as list of lists of data frame.}

\item{dims}{An n-dimensional integer vector. Each element of this vector
indicates the size of the corresponding dimension in the n-dimensional
matrixc.}
}
\value{
In integer vector with linear indexes.
}
\description{
Takes subscripts as list of lists or data frame and converts them to linear
index.
}
\keyword{internal}
