% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{Ark}
\alias{Ark}
\title{A Class to represent an ark.}
\description{
An Ark object can create unique pseudonyms.
Given the same input, it will always return the same psuedonym.
No pseudonym will repeat.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{log}}{Hashtable for all used pseudonyms. Inputs (keys) are stored
as hashes.}

\item{\code{parts}}{Words that will be combined to form pseudonyms.}

\item{\code{max_length}}{Maximum number of possible pseudonyms in the Ark.}

\item{\code{index_shuffled}}{a random permutation of the index}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parts}}{Words that will be combined to form pseudonyms.}

\item{\code{max_length}}{Maximum number of possible pseudonyms in the Ark.}

\item{\code{index_shuffled}}{a random permutation of the index}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ark$new()}}
\item \href{#method-pseudonymize}{\code{Ark$pseudonymize()}}
\item \href{#method-print}{\code{Ark$print()}}
\item \href{#method-length}{\code{Ark$length()}}
\item \href{#method-clone}{\code{Ark$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new arc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Ark} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pseudonymize"></a>}}
\if{latex}{\out{\hypertarget{method-pseudonymize}{}}}
\subsection{Method \code{pseudonymize()}}{
Create Pseudonyms for input.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$pseudonymize(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{One or more R objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector of pseudonyms with same length as input.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Pretty-print an Ark object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\if{latex}{\out{\hypertarget{method-length}{}}}
\subsection{Method \code{length()}}{
Number of used pseudonyms in an Ark.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$length(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other arguments passed on to \code{length()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ark$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
