% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noah.R
\name{pseudonymize}
\alias{pseudonymize}
\title{Create unique pseudonyms.}
\usage{
pseudonymize(..., .ark = NULL)
}
\arguments{
\item{...}{One or multiple objects for which pseudonyms should be created,
usually one or more columns of a data frame. All objects must be of the same
length.}

\item{.ark}{An Ark object. By default a new Ark is created. Using an existing
Ark makes sure that the same input returns the same pseudonym.}
}
\value{
A character vector of pseudonyms.
}
\description{
Pseudonymize returns unique pseudonyms for R objects
}
\examples{
pseudonymize("Mata Hari")
}
